Fix download.sh

+++ download.sh
@@ -7,29 +7,45 @@ function clean() {
 	rm -Rf $TOCLEAN
 	TOCLEAN=""
 }
+trap clean EXIT
+
+function mktempd() {
+	d="$(mktemp -d)"
+	TOCLEAN+="$d"
+}
 
 find known-imgs -type f |while read i;do
 	folder="$(cut -d / -f 2- <<<$i)"
 	mkdir -p output/$folder
 	destination="output/$folder/orig-boot.img"
+	if [ -f "$destination" ];then
+		continue
+	fi
 
 	curr=""
-	cat $i |while read j;do
+	OLD_IFS="$IFS"
+	IFS=$'\n'
+	for j in $(cat $i);do
 		if grep -qE '^http' <<<$j;then
-			curr="$(mktemp)"
+			mktempd
+			curr="$d/$(basename "$j")"
 			wget "$j" -O $curr
 		elif grep -qE '\.zip$' <<<$curr;then
-			d="$(mktemp -d)"
-			TOCLEAN="$TOCLEAN $d"
-			unzip "$curr" "$j" -d "$d"
+			mktempd
+			unzip "$curr" "$j" -d "$d" ||	\
+				unzip "$curr" "*/$j" -d "$d"
+
 			curr="$(find "$d" -name "$(basename "$j")")"
 		elif grep -qE '\.tgz$' <<<$curr;then
-			d="$(mktemp -d)"
-			TOCLEAN="$TOCLEAN $d"
-			tar xf "$curr" "$j" -C "$d"
+			mktempd
+			file="$(tar tf "$curr" |grep -E "/$j")"
+			tar xf "$curr" -C "$d" "$file"
 			curr="$(find "$d" -name "$(basename "$j")")"
 		fi
+		echo $curr
 	done
-	cp "$curr" "$destination"
+	IFS="$OLD_IFS"
+	cp "${curr}" "$destination"
+	rm -f result
 	clean
 done
